separate token generation and logic out into its own service class honering SRP
