+++ lineshift-v3-backend/Services/Identity/AuthServices.cs
@@ -27,24 +27,24 @@ namespace lineshift_v3_backend.Services.Identity
         private readonly ILogger<AuthServices> _logger;
         private readonly UserManager<ApplicationUser> _userManager;
         private readonly SignInManager<ApplicationUser> _signInManager;
-        private readonly IConfiguration _configuration;
+        private readonly ITokenServices _tokenServices;
 
         public AuthServices(
             IAuthRepository authRepository, 
             ILogger<AuthServices> logger, 
             UserManager<ApplicationUser> userManager,
             SignInManager<ApplicationUser> signInManager,
-            IConfiguration configuration)
+            ITokenServices tokenServices)
         {
             _authRepository = authRepository;
             _logger = logger;
             _userManager = userManager;
             _signInManager = signInManager;
-            _configuration = configuration;
+            _tokenServices = tokenServices;
 
         }
 
-        #region Service Methods
+
         public async Task<Result<AuthResponse>> LoginAsync(LoginModel loginModel)
         {
             var userEntity = await _authRepository.LoginAsync(loginModel);
@@ -80,7 +80,7 @@ namespace lineshift_v3_backend.Services.Identity
             await _userManager.UpdateAsync(userEntity);
 
             // Generate JWT Token
-            var jwt_token = await GenerateJwtToken(userEntity);
+            var jwt_token = await _tokenServices.GenerateJwtToken(userEntity);
 
             // Making Session User
             var sessionUser = await MapSessionUser(userEntity);
@@ -151,50 +151,5 @@ namespace lineshift_v3_backend.Services.Identity
             }
         }
         #endregion
-
-        internal async Task<string> GenerateJwtToken(ApplicationUser userEntity)
-        {
-            var claims = new List<Claim>
-            {
-                new Claim(JwtRegisteredClaimNames.Sub, userEntity.Id),
-                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()), // Unique Id for the token
-                new Claim(ClaimTypes.NameIdentifier, userEntity.Id), // Standard Claim for User ID
-                new Claim(ClaimTypes.Name, userEntity.UserName ?? userEntity.UserName ?? ""), // Standard Claim for Username
-                new Claim(ClaimTypes.Email, userEntity.Email ?? ""), // Standard Claim for User Email
-            };
-
-            var roles = await _userManager.GetRolesAsync(userEntity);
-            foreach (var role in roles)
-            {
-                claims.Add(new Claim(ClaimTypes.Role, role));
-            }
-
-            // Any Additional custom claims for ApplicationUser Model 
-            // EX: if we wanted the users FirstName and LastName for display purposes
-
-            // Get JWT config from appsetting.json
-            var jwtKey = _configuration["Jwt:Key"] ?? throw new InvalidOperationException("JWT Key not found in configuration.");
-            var jwtIssuer = _configuration["Jwt:Issuer"] ?? throw new InvalidOperationException("JWT Issuer not found in configuration.");
-            var jwtAudience = _configuration["Jwt:Audience"] ?? throw new InvalidOperationException("JWT Audience not found in configuration.");
-            var expireDays = Convert.ToDouble(_configuration["Jwt:ExpireDays"] ?? "7");
-
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtKey));
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-            var expires = DateTime.Now.AddDays(expireDays);
-
-            var token = new JwtSecurityToken(
-                issuer: jwtIssuer,
-                audience: jwtAudience,
-                claims: claims,
-                expires: expires,
-                signingCredentials: creds
-            );
-
-            return new JwtSecurityTokenHandler().WriteToken(token);
-        }
-        #endregion
-
-
-        
     }
 }
