+++ lineshift-v3-backend/Services/Identity/TokenServices.cs
@@ -1,0 +1,73 @@
+﻿using lineshift_v3_backend.Models.Database;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+
+namespace lineshift_v3_backend.Services.Identity
+{
+    #region Token Service Interface
+    public interface ITokenServices
+    {
+        Task<string> GenerateJwtToken(ApplicationUser userEntity);
+    }
+    #endregion
+
+    public class TokenServices : ITokenServices
+    {
+        private readonly IConfiguration _configuration;
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        public TokenServices 
+        (
+            IConfiguration configuration, 
+            UserManager<ApplicationUser> userManager)
+        {
+            _configuration = configuration;
+            _userManager = userManager;
+        }
+
+        public async Task<string> GenerateJwtToken(ApplicationUser userEntity)
+        {
+            var claims = new List<Claim>
+            {
+                new Claim(JwtRegisteredClaimNames.Sub, userEntity.Id),
+                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()), // Unique Id for the token
+                new Claim(ClaimTypes.NameIdentifier, userEntity.Id), // Standard Claim for User ID
+                new Claim(ClaimTypes.Name, userEntity.UserName ?? userEntity.UserName ?? ""), // Standard Claim for Username
+                new Claim(ClaimTypes.Email, userEntity.Email ?? ""), // Standard Claim for User Email
+            };
+
+            var roles = await _userManager.GetRolesAsync(userEntity);
+            foreach (var role in roles)
+            {
+                claims.Add(new Claim(ClaimTypes.Role, role));
+            }
+
+            // Any Additional custom claims for ApplicationUser Model 
+            // EX: if we wanted the users FirstName and LastName for display purposes
+
+            // Get JWT config from appsetting.json
+            var jwtKey = _configuration["Jwt:Key"] ?? throw new InvalidOperationException("JWT Key not found in configuration.");
+            var jwtIssuer = _configuration["Jwt:Issuer"] ?? throw new InvalidOperationException("JWT Issuer not found in configuration.");
+            var jwtAudience = _configuration["Jwt:Audience"] ?? throw new InvalidOperationException("JWT Audience not found in configuration.");
+            var expireDays = Convert.ToDouble(_configuration["Jwt:ExpireDays"] ?? "7");
+
+            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtKey));
+            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
+            var expires = DateTime.Now.AddDays(expireDays);
+
+            var token = new JwtSecurityToken(
+                issuer: jwtIssuer,
+                audience: jwtAudience,
+                claims: claims,
+                expires: expires,
+                signingCredentials: creds
+            );
+
+            return new JwtSecurityTokenHandler().WriteToken(token);
+        }
+    }
+
+}
